Apply Review Suggestion
